requirements.txt
langchain~=0.3.25
selenium~=4.33.0
langgraph~=0.4.8
python-dotenv~=1.1.0
streamlit~=1.45.1
pandas~=2.3.0
plotly~=6.1.2
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.40
selenium>=4.15.0
webdriver-manager>=4.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
urllib3>=2.0.0